# winalog-go/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/dashboard": {
            "get": {
                "description": "初始化DashboardHandler",
                "tags": [
                    "dashboard"
                ],
                "summary": "创建仪表板处理器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据库实例",
                        "name": "db",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/ui": {
            "get": {
                "description": "初始化UIHandler",
                "tags": [
                    "ui"
                ],
                "summary": "创建UI处理器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据库实例",
                        "name": "db",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}